refactor core manager into main file and move global types to a regular ts file
