+++ src/managers/manager.ts
