--- src/types/global.d.ts
@@ -1,6 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import _ from 'lodash';
-
-declare global {
-  const _: typeof _;
-}
